+++ test/dummy/app/uchi/repositories/book.rb
@@ -3,7 +3,10 @@ module Uchi
     class Book < Repository
       def fields
         [
-          Field::HasMany.new(:titles),
+          Field::HasMany.new(:titles).nested_fields(
+            Field::String.new(:title),
+            Field::String.new(:locale)
+          ),
           Field::String.new(:original_title)
         ]
       end